fix database query error bug
